refactor: controller 이름 변경
